Added changes for Tensorflow-GPU, folder paths, Path Variable & dir_path arg of visualize_explanation()
Updated the code to train the model on Tensorflow-GPU, previously it just ran on tensorflow CPU.
Updated folder paths (from \\ to /). The \\ creates a new folder with preceding labels(parent folder or filename) as name of a single folder on Linux.
Add code for Path Variable & updated dir_path arg of visualize_explanation().

The above code works on Linux, haven't checked it on windows.